フェイズマネージャーwip
