--- Assets/1Ito/Novel/Scripts/NovelManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43d732a7c3e194344b1f21b782165ebd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
